+++ libs/common/src/lib.rs
@@ -3,6 +3,7 @@ pub mod graphql;
 pub mod http;
 pub mod json;
 pub mod plugin;
+pub mod plugin_manager;
 pub mod serde_utils;
 pub mod vrl_functions;
 pub mod vrl_utils;
